+++ pkg/providers/property/testdata/TestDataEdgeHostname/missing_group_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_edge_hostname" "test" {
+  id          = "ehn_123"
+  contract_id = "ctr_123"
+}